batch/covid19_hashtag/requirement.txt
pandas==1.0.3
plotly==4.6.0
boto3==1.6.11
chart_studio==1.1.0
Pillow==7.0.0